# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: StorageHub API
  version: 1.0.0
  description: >
    REST API of the StorageHub self-service storage management system.
    All endpoints live under the `/api/v1` prefix. Everything except the
    explicitly public endpoints requires a JWT Bearer token (AD-5).
servers:
  - url: /api/v1
    description: Same-origin (Vite dev proxy / SPA-served deployment)

# Default security: JWT everywhere. Public endpoints (e.g. POST /auth/login)
# override this with `security: []` when they are added.
security:
  - bearerAuth: []
paths:
...
  /reservations:
    post:
      operationId: createReservation
      tags: [Reservations]
      summary: Create a pending reservation with locked price snapshot
      description: >
        Atomically validates unit availability and creates a PENDING_PAYMENT reservation
        with a unique BK- code and immutable price snapshot. Returns 409 UNIT_UNAVAILABLE
        if the unit is occupied, under maintenance, or has conflicting reservations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation created successfully in PENDING_PAYMENT status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Unit is unavailable or conflicting reservation exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reservations/my:
    get:
      operationId: getMyReservations
      tags: [Reservations]
      summary: List reservations belonging to the authenticated customer
      description: >
        Returns list of reservations and rentals created by the authenticated customer,
        ordered by most recently created first.
      responses:
        '200':
          description: List of reservations belonging to customer.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [Reservations]
      summary: Get reservation details by ID
      description: >
        Returns reservation details including unit info, timeline, status, and price snapshot.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reservation ID.
      responses:
        '200':
          description: Reservation details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /reservations/{id}/access-code:
    get:
      operationId: getReservationAccessCode
      tags: [Reservations]
      summary: Reveal access code for checked-in rental (SensitiveValue AD-5)
      description: >
        Guarded endpoint to reveal sensitive access code (PIN/QR). Accessible by customer of reservation or staff/admin
        only after check-in ritual has completed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reservation ID.
      responses:
        '200':
          description: Sensitive access code revealed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessCodeResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /reservations/{id}/extension-boundary:
    get:
      operationId: getExtensionBoundary
      tags: [Reservations]
      summary: Get extension conflict boundary and latest allowed checkout date (Story 4.1)
      description: >
        Returns conflict boundary, next upcoming reservation dates, and max extendable checkout date
        for an active CHECKED_IN rental.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reservation ID.
      responses:
        '200':
          description: Extension conflict boundary details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtensionBoundaryDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /reservations/{id}/extension-quote:
    post:
      operationId: getExtensionQuote
      tags: [Reservations]
      summary: Calculate itemized rental extension fee and deposit top-up (Story 4.1 & 4.2)
      description: >
        Validates target checkout date against conflict boundary and computes 2-line financial breakdown
        (additional rent based on locked rate + deposit top-up formula AD-11).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reservation ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtensionQuoteRequest'
      responses:
        '200':
          description: Itemized extension fee breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtensionQuoteDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Target extension date conflicts with an existing reservation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------------------------------------------- payments (2.4)
...
  /contracts/reservation/{reservationId}:
    get:
      operationId: getContractByReservation
      tags: [Contracts]
      summary: Get latest contract by reservation ID
      description: >
        Returns the latest active contract associated with a reservation.
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reservation ID.
      responses:
        '200':
          description: Latest contract for reservation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT with role claim, 24h TTL, no refresh token (AD-5). Logout is the
        client dropping the token.

  # D6 reusable non-2xx bodies (deferred from 1.1, landed with the first
  # auth paths in 1.3): every usage renders the same Error envelope.
  responses:
    NotFound:
      description: >
        Resource was not found. Error envelope, code NOT_FOUND (404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: >
        Missing, invalid or expired Bearer token - or bad credentials /
        inactive / locked account at login. Error envelope, code
        UNAUTHENTICATED (401). Carries `WWW-Authenticate: Bearer` on guarded
        paths. At login every failure branch shares one generic message.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >
        Authenticated but the token's role is not allowed on this endpoint
        (server-side permission matrix). Error envelope, code FORBIDDEN (403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: >
        Request body failed validation (missing/invalid fields, duplicate
        email, password confirmation mismatch, terms not accepted). Error
        envelope, code VALIDATION_FAILED (400) with one fieldError per
        offending field.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      description: >
        The one error envelope for every non-2xx response, including security
        (401/403) and business-rule blocks (409). The FE renders toasts from
        this data and never parses the human message.
      required:
        - code
        - message
      additionalProperties: false
      properties:
        code:
          type: string
          description: >
            Machine-readable, stable, UPPER_SNAKE. Two layers share this field.
            Framework codes, issued by the backend error envelope for every
            non-2xx response including security 401/403: VALIDATION_FAILED
            (400, with fieldErrors), MALFORMED_REQUEST (400), UNAUTHENTICATED
            (401), FORBIDDEN (403), NOT_FOUND (404), METHOD_NOT_ALLOWED (405),
            UNSUPPORTED_MEDIA_TYPE (415), INTERNAL_ERROR (500, generic message
            only). Business codes are declared by each operation when its path
            lands in this file and travel in this same field; a business-rule
            block maps to 409.
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples:
            - VALIDATION_FAILED
            - UNIT_UNAVAILABLE
            - BLOCKED_TRANSITION
        message:
          type: string
          description: Human-readable one-liner (what happened + consequence + next step).
        fieldErrors:
          type: array
          description: Present for validation failures; empty or absent otherwise.
          items:
            $ref: '#/components/schemas/FieldError'
        missingStep:
          type: string
          nullable: true
          description: Machine code for the missing closing step (FR-22 snap-back).
          example: 'RENT_PAYMENT_PENDING'
        stepLabel:
          type: string
          nullable: true
          description: Human-readable closing step name for rendering the snap-back toast.
          example: 'Rent payment is still pending on this check-in.'
    FieldError:
      type: object
      required:
        - field
        - message
      additionalProperties: false
      properties:
        field:
          type: string
          description: camelCase JSON path of the offending field.
        message:
          type: string

    # ------------------------------------------------------------------- list
...
    CreateReservationRequest:
      type: object
      description: Payload to atomically validate and create a pending reservation.
      required:
        - unitCode
        - startDate
        - durationMonths
      additionalProperties: false
      properties:
        unitCode:
          type: string
          description: Unique unit code (e.g. S-3, M-5).
          example: 'M-5'
        startDate:
          type: string
          format: date
          description: Desired rental start date (YYYY-MM-DD).
          example: '2026-10-03'
        durationMonths:
          type: integer
          minimum: 1
          maximum: 120
          description: Rental duration in months.
          example: 1
    ReservationStatus:
      type: string
      description: Lifecycle status of a reservation.
      enum:
        - PENDING_PAYMENT
        - RESERVED
        - CHECKED_IN
        - CHECKOUT_REQUESTED
        - CLOSED
        - EXPIRED
        - CANCELLED
    ReservationDto:
      type: object
      description: Full reservation representation including locked price snapshot.
      required:
        - id
        - code
        - customerId
        - unitId
        - unitCode
        - startDate
        - endDate
        - durationMonths
        - depositAmount
        - monthlyRate
        - baseRent
        - totalRent
        - policyVersion
        - status
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          example: 1
        code:
          type: string
          example: 'BK-2026-0001'
        customerId:
          type: integer
          format: int64
          example: 1
        customerName:
          type: string
          nullable: true
          example: 'Lan Nguyen'
        unitId:
          type: integer
          format: int64
          example: 3
        unitCode:
          type: string
          example: 'M-5'
        unitTypeName:
          type: string
          nullable: true
          example: 'M'
        facilityName:
          type: string
          nullable: true
          example: 'Tan Binh Depot'
        facilityAddress:
          type: string
          nullable: true
          example: '45 Nguyen Van Troi, Tan Binh, Ho Chi Minh City'
        zoneCode:
          type: string
          nullable: true
          example: 'B'
        floor:
          type: integer
          nullable: true
          example: 2
        sizeM2:
          type: number
          format: double
          nullable: true
          example: 8.0
        accessType:
          type: string
          nullable: true
          example: 'QR'
        startDate:
          type: string
          format: date
          example: '2026-10-03'
        endDate:
          type: string
          format: date
          example: '2026-11-03'
        durationMonths:
          type: integer
          example: 1
        depositAmount:
          type: integer
          description: Locked deposit amount in VND.
          example: 69000
        monthlyRate:
          type: integer
          description: Locked monthly rate in VND.
          example: 690000
        baseRent:
          type: integer
          description: Locked base rent in VND (monthlyRate * durationMonths).
          example: 690000
        totalRent:
          type: integer
          description: Locked total rent in VND.
          example: 690000
        policyVersion:
          type: string
          description: Locked rental policy version.
          example: 'v3'
        accessCode:
          type: string
          nullable: true
          example: '482913'
        status:
          $ref: '#/components/schemas/ReservationStatus'
        payments:
          type: array
          description: Associated payment receipts and transactions.
          items:
            $ref: '#/components/schemas/PaymentDto'

    # ----------------------------------------------------------- payments (2.4)
    PaymentStatus:
      type: string
      description: Lifecycle status of a payment.
      enum:
        - PENDING
        - PENDING_CASH
        - PROCESSING
        - SUCCEEDED
        - FAILED
        - EXPIRED
    PaymentPurpose:
      type: string
      description: Purpose of payment transaction.
      enum:
        - DEPOSIT
        - RENT
        - EXTENSION_FEE
        - DAMAGE_FEE
        - EXTRA_FEE
    PaymentMethod:
      type: string
      description: Payment method.
      enum:
        - PAYOS
        - CASH
        - CARD
        - MOMO
        - VNPAY
...
    PaymentDto:
      type: object
      description: Full payment detail.
      required:
        - id
        - receiptCode
        - payerId
        - orderCode
        - purpose
        - method
        - amount
        - status
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          example: 10
        receiptCode:
          type: string
          example: 'RC-2026-0001'
        payerId:
          type: integer
          format: int64
          example: 1
        reservationId:
          type: integer
          format: int64
          nullable: true
          example: 1
        orderCode:
          type: integer
          format: int64
          example: 1727932800101
        purpose:
          $ref: '#/components/schemas/PaymentPurpose'
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount:
          type: integer
          example: 69000
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
          nullable: true
...
    ContractStatus:
      type: string
      description: Lifecycle status of a contract.
      enum:
        - DRAFT
        - PRINTED
        - SIGNED
        - ACTIVE
        - CLOSED
        - SUPERSEDED
    ContractContentSnapshot:
      type: object
      description: Immutable locked terms snapshot recorded at contract creation.
      required:
        - code
        - reservationCode
        - unitCode
        - monthlyRate
        - baseRent
        - totalRent
        - depositAmount
        - depositRate
        - durationMonths
        - policyVersion
        - currency
      additionalProperties: true
      properties:
        code:
          type: string
          example: 'CT-2026-0001'
        reservationCode:
          type: string
          example: 'BK-2026-0001'
        unitCode:
          type: string
          example: 'M-5'
        monthlyRate:
          type: integer
          example: 690000
        baseRent:
          type: integer
          example: 690000
        totalRent:
          type: integer
          example: 690000
        depositAmount:
          type: integer
          example: 69000
        depositRate:
          type: integer
          example: 10
        durationMonths:
          type: integer
          example: 1
        policyVersion:
          type: string
          example: 'v3'
        currency:
          type: string
          example: 'VND'
        startDate:
          type: string
          format: date
          example: '2026-10-03'
        endDate:
          type: string
          format: date
          example: '2026-11-03'
        customerName:
          type: string
          nullable: true
          example: 'Lan Nguyen'
        customerEmail:
          type: string
          nullable: true
          example: 'lan.nguyen@example.com'
        customerPhone:
          type: string
          nullable: true
          example: '0901234567'
        facilityName:
          type: string
          nullable: true
          example: 'Tan Binh Depot'
        facilityAddress:
          type: string
          nullable: true
          example: '45 Nguyen Van Troi, Tan Binh, Ho Chi Minh City'
        zoneCode:
          type: string
          nullable: true
          example: 'B'
        floor:
          type: integer
          nullable: true
          example: 2
        sizeM2:
          type: number
          format: double
          nullable: true
          example: 8.0
    ContractDto:
      type: object
      description: Full contract representation including locked snapshot terms and revision chaining.
      required:
        - id
        - code
        - reservationId
        - status
        - isLatest
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          example: 1
        code:
          type: string
          example: 'CT-2026-0001'
        reservationId:
          type: integer
          format: int64
          example: 1
        reservationCode:
          type: string
          nullable: true
          example: 'BK-2026-0001'
        policyId:
          type: integer
          nullable: true
          example: 1
        policyVersion:
          type: string
          nullable: true
          example: 'v3'
        contentSnapshot:
          type: string
          description: Raw JSON string of immutable terms snapshot.
        snapshot:
          $ref: '#/components/schemas/ContractContentSnapshot'
        signedPhotoUrl:
          type: string
          nullable: true
          example: '/api/v1/attachments/signed-ct-1.jpg'
        status:
          $ref: '#/components/schemas/ContractStatus'
        supersedesContractId:
          type: integer
          format: int64
          nullable: true
        isLatest:
          type: integer
          example: 1

    # ----------------------------------------------------------------- tasks (3.2)
...
    AccessCodeResponseDto:
      type: object
      description: Sensitive access code response for checked-in rental (SensitiveValue AD-5).
      required:
        - accessCode
        - accessType
        - unitCode
      additionalProperties: false
      properties:
        accessCode:
          type: string
          example: '482913'
        accessType:
          type: string
          example: 'PIN'
        unitCode:
          type: string
          example: 'S-3'

    # -------------------------------------------------------- extensions (4.1, 4.2)
    ExtensionBoundaryDto:
      type: object
      description: Conflict boundary and allowed checkout window for rental extension (Story 4.1).
      required:
        - rentalId
        - unitCode
        - currentEndDate
        - isExtendable
      additionalProperties: false
      properties:
        rentalId:
          type: integer
          format: int64
          example: 1
        unitCode:
          type: string
          example: 'M-2'
        currentEndDate:
          type: string
          format: date
          example: '2026-10-15'
        latestPossibleCheckoutDate:
          type: string
          format: date
          nullable: true
          description: Max checkout date before next upcoming booking.
          example: '2026-10-18'
        conflictStartDate:
          type: string
          format: date
          nullable: true
          description: Start date of the next conflicting reservation.
          example: '2026-10-19'
        conflictReservationCode:
          type: string
          nullable: true
          example: 'BK-2026-0045'
        isExtendable:
          type: boolean
          description: Whether the rental is CHECKED_IN and not past endDate.
          example: true
        message:
          type: string
          nullable: true
          example: 'M-2 has a reservation starting Oct 19. Latest possible checkout is Oct 18.'
    ExtensionQuoteRequest:
      type: object
      required:
        - newEndDate
      additionalProperties: false
      properties:
        newEndDate:
          type: string
          format: date
          example: '2026-10-18'
    ExtensionQuoteDto:
      type: object
      description: Itemized rental extension fee calculation result.
      required:
        - rentalId
        - unitCode
        - currentEndDate
        - newEndDate
        - additionalDays
        - monthlyRate
        - additionalRent
        - currentHeldDeposit
        - newTotalDepositRequired
        - depositTopUp
        - totalFee
        - currency
      additionalProperties: false
      properties:
        rentalId:
          type: integer
          format: int64
          example: 1
        unitCode:
          type: string
          example: 'M-2'
        currentEndDate:
          type: string
          format: date
          example: '2026-10-15'
        newEndDate:
          type: string
          format: date
          example: '2026-10-18'
        additionalDays:
          type: integer
          example: 3
        additionalMonths:
          type: number
          format: double
          example: 0.1
        monthlyRate:
          type: integer
          format: int64
          example: 690000
        additionalRent:
          type: integer
          format: int64
          example: 690000
        currentHeldDeposit:
          type: integer
          format: int64
          example: 103500
        newTotalDepositRequired:
          type: integer
          format: int64
          example: 172500
        depositTopUp:
          type: integer
          format: int64
          description: Top-up deposit required = max(0, newDeposit - currentHeldDeposit)
          example: 69000
        totalFee:
          type: integer
          format: int64
          description: Total payable fee = additionalRent + depositTopUp
          example: 759000
        currency:
          type: string
          example: 'VND'
        policyVersion:
          type: string
          nullable: true
          example: 'v3'